split branch/leaf from rrcf
