+++ rrcf/leaf.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+
+class Leaf:
+    """
+    Leaf of RCTree containing no children and at most one parent.
+
+    Attributes:
+    -----------
+    i: Index of leaf (user-specified)
+    d: Depth of leaf
+    u: Pointer to parent
+    x: Original point (1 x d)
+    n: Number of points in leaf (1 if no duplicates)
+    b: Bounding box of point (1 x d)
+    """
+    __slots__ = ['i', 'd', 'u', 'x', 'n', 'b']
+
+    def __init__(self, i, d=None, u=None, x=None, n=1):
+        self.u = u
+        self.i = i
+        self.d = d
+        self.x = x
+        self.n = n
+        self.b = x.reshape(1, -1)
+
+    def __repr__(self):
+        return "Leaf({0})".format(self.i)
